Relocate auth0_logout_url method to LogoutController
It is only used there.
